implement Cloneable and extend Character
